% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data_directory}
\alias{data_directory}
\title{Verzeichnis von Datensätzen}
\usage{
data_directory()
}
\value{
Ein Verzeichnis der verfügbaren Daten. Bei Nutzung im Stadtnetz sind auch (noch) nicht veröffentlichte Daten verfügbar.
add(1, 1)
}
\description{
Verzeichnis von Datensätzen
}
