% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{uster_data_directory}
\alias{uster_data_directory}
\title{Verzeichnis von Datensätzen}
\usage{
uster_data_directory()
}
\value{
Ein Verzeichnis der Verfügbaren Daten. Bei stadtinterner Nutzung sind auch (noch) nicht öffentliche Daten verfügbar.
add(1, 1)
}
\description{
Verzeichnis von Datensätzen
}
