% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{usteR_load_color}
\alias{usteR_load_color}
\title{Farb-Schemas gemäss CD der Stadt Uster.}
\usage{
usteR_load_color()
}
\arguments{
\item{schema}{Die Bezeichnung des Schemas:}
}
\value{
Das Farbschema als Vector
add(1, 1)
}
\description{
Farb-Schemas gemäss CD der Stadt Uster.
}
