% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{usteR_load_data}
\alias{usteR_load_data}
\title{Laden eines Datensatz anhand der ID}
\usage{
usteR_load_data(id)
}
\arguments{
\item{id}{ID des gewünschten Datensatz}
}
\value{
Ein Dataset
add(1, 1)
}
\description{
Laden eines Datensatz anhand der ID
}
