% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{usteR_data_directory}
\alias{usteR_data_directory}
\title{Verzeichnis von Datensätzen}
\usage{
usteR_data_directory()
}
\value{
Ein Verzeichnis der Verfügbaren Daten. Bei interner Nutzung in der Stadt sind auch (noch) nicht veröffentlichte Daten verfügbare.
add(1, 1)
}
\description{
Verzeichnis von Datensätzen
}
