% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{uster_load_data}
\alias{uster_load_data}
\title{Laden eines Datensatz anhand der ID}
\usage{
uster_load_data(id)
}
\arguments{
\item{ID}{ID des gewünschten Datensatz}
}
\value{
Die Summe von \code{x} und \code{y}. #' @examples
add(1, 1)
}
\description{
Laden eines Datensatz anhand der ID
}
